+++ pytorchrl/agent/storages/on_policy/ppod_buffer.py
@@ -60,8 +60,8 @@ class PPODBuffer(B):
 
     def __init__(self, size, device, actor, algorithm, envs, rho=0.1, phi=0.3, gae_lambda=0.95, alpha=10,
                  total_buffer_demo_capacity=51, initial_human_demos_dir=None, initial_agent_demos_dir=None,
-                 target_agent_demos_dir=None, save_demos_every=10, num_agent_demos_to_save=10,
-                 initial_reward_threshold=None):
+                 target_agent_demos_dir=None,  num_agent_demos_to_save=10, initial_reward_threshold=None,
+                 save_demos_every=10, demo_dtypes={prl.OBS: np.uint8, prl.ACT: np.int8,  prl.REW: np.float16}):
 
         super(PPODBuffer, self).__init__(
             size=size,
@@ -87,10 +87,9 @@ class PPODBuffer(B):
         self.target_agent_demos_dir = target_agent_demos_dir
 
         # Data parameters
-        self.demo_act_dtype = np.float32
-        self.demo_obs_dtype = np.float32
-        self.demo_rew_dtype = np.float32
+        self.demo_dtypes = demo_dtypes
         self.num_channels_obs = None  # Lazy initialization
+        self.inserted_samples = 0
 
         # Reward and Value buffers
         self.reward_demos = []
@@ -120,7 +119,7 @@ class PPODBuffer(B):
     @classmethod
     def create_factory(cls, size, rho=0.1, phi=0.3, gae_lambda=0.95, alpha=10, total_buffer_demo_capacity=51,
                        initial_human_demos_dir=None, initial_agent_demos_dir=None, target_agent_demos_dir=None,
-                       save_demos_every=10, num_agent_demos_to_save=10, initial_reward_threshold=None):
+                       num_agent_demos_to_save=10, initial_reward_threshold=None, save_demos_every=10):
         """
         Returns a function that creates PPODBuffer instances.
 
@@ -160,8 +159,8 @@ class PPODBuffer(B):
         def create_buffer_instance(device, actor, algorithm, envs):
             """Create and return a PPODBuffer instance."""
             return cls(size, device, actor, algorithm, envs, rho, phi, gae_lambda, alpha, total_buffer_demo_capacity,
-                       initial_human_demos_dir, initial_agent_demos_dir, target_agent_demos_dir, save_demos_every,
-                       num_agent_demos_to_save, initial_reward_threshold)
+                       initial_human_demos_dir, initial_agent_demos_dir, target_agent_demos_dir,
+                       num_agent_demos_to_save, initial_reward_threshold, save_demos_every)
         return create_buffer_instance
 
     def before_gradients(self):
@@ -192,7 +191,7 @@ class PPODBuffer(B):
             for x in self.data[prl.RHS]:
                 self.data[prl.RHS][x][step].copy_(next_rhs[x])
         else:
-            self.data[prl.RHS][step].copy_(next_rhs)
+            self.data[prl.RHS][step] = next_rhs
 
         # Compute returns and advantages
         if isinstance(self.data[prl.VAL], dict):
@@ -209,13 +208,63 @@ class PPODBuffer(B):
                 self.data[prl.REW], self.data[prl.RET], self.data[prl.VAL], self.data[prl.DONE], self.algo.gamma)
             self.data[prl.ADV] = self.compute_advantages(self.data[prl.RET], self.data[prl.VAL])
 
-        # self.data[prl.VAL][step].copy_(value_dict.get("value_net1"))
-        # self.compute_returns()
-        # self.compute_advantages()
+        if hasattr(self.algo, "gamma_intrinsic") and prl.IREW in self.data.keys():
+            self.normalize_int_rewards()
+            self.algo.state_rms.update(
+                self.data[prl.OBS][:, :, -self.num_channels_obs:, ...].reshape(-1, 1, *self.data[prl.OBS].shape[3:]))
+
+        # If algorithm with intrinsic rewards, also compute ireturns and iadvantages
+        if prl.IVAL in self.data.keys() and prl.IREW in self.data.keys():
+            if isinstance(self.data[prl.IVAL], dict):
+                # If multiple critics
+                for x in self.data[prl.IVAL]:
+                    self.data[prl.IVAL][x][step].copy_(value_dict.get(x))
+                    self.compute_returns(
+                        self.data[prl.IREW], self.data[prl.IRET][x], self.data[prl.IVAL][x],
+                        torch.zeros_like(self.data[prl.DONE]), self.algo.gamma_intrinsic)
+                    self.data[prl.IADV][x] = self.compute_advantages(self.data[prl.IRET][x], self.data[prl.IVAL][x])
+            else:
+                # If single critic
+                self.data[prl.IVAL][step].copy_(value_dict.get("ivalue_net1"))
+                self.compute_returns(
+                    self.data[prl.IREW], self.data[prl.IRET], self.data[prl.IVAL],
+                    torch.zeros_like(self.data[prl.DONE]), self.algo.gamma_intrinsic)
+                self.data[prl.IADV] = self.compute_advantages(self.data[prl.IRET], self.data[prl.IVAL])
 
         self.iter += 1
         if self.iter % self.save_demos_every == 0:
             self.save_demos()
+
+    def after_gradients(self, batch, info):
+        """
+        After updating actor policy model, make sure self.step is at 0.
+        Parameters
+        ----------
+        batch : dict
+            Data batch used to compute the gradients.
+        info : dict
+            Additional relevant info from gradient computation.
+        Returns
+        -------
+        info : dict
+            info dict updated with relevant info from Storage.
+        """
+
+        step = self.step if self.step != 0 else -1
+        for k in (prl.OBS, prl.RHS, prl.DONE):
+            if isinstance(self.data[k], dict):
+                for x in self.data[k]:
+                    self.data[k][x][0].copy_(self.data[k][x][step])
+            else:
+                self.data[k][0].copy_(self.data[k][step])
+
+        if self.step != 0:
+            self.step = 0
+
+        # info['NumberSamples'] -= self.inserted_samples
+        self.inserted_samples = 0
+
+        return info
 
     def get_num_channels_obs(self, sample):
         """
@@ -272,14 +321,11 @@ class PPODBuffer(B):
         # Track episodes for potential demos
         self.track_potential_demos(sample)
 
-        # Handle demos in progress
+        all_envs, all_obs, all_done, all_rhs = [], [], [], {k: [] for k in sample[prl.RHS].keys()}
+
         for i in range(self.num_envs):
 
-            # For each environment, insert demo transition if in the middle of a demos
             if self.demos_in_progress["env{}".format(i + 1)]["Demo"]:
-
-                # Get next demo step to be inserted
-                demo_step = self.demos_in_progress["env{}".format(i + 1)]["Step"]
 
                 # Get demo obs, rhs and done tensors to run forward pass
                 obs = self.data[prl.OBS][self.step][i].unsqueeze(0)
@@ -289,29 +335,55 @@ class PPODBuffer(B):
                 else:
                     obs, rhs, done = self.actor.actor_initial_states(obs)
 
-                # Run forward pass
-                _, _, rhs2, algo_data = self.algo.acting_step(obs, rhs, done)
+                all_envs.append(i)
+                all_obs.append(obs)
+                all_done.append(done)
+                for k in rhs:
+                    all_rhs[k].append(rhs[k])
 
-                # Insert demo act tensor to self.step
-                self.data[prl.ACT][self.step][i].copy_(self.demos_in_progress["env{}".format(
-                    i + 1)]["Demo"][prl.ACT][demo_step])
+            # Otherwise check if end of episode reached and randomly start new demo
+            elif sample[prl.DONE2][i] == 1.0:
 
-                # Insert demo rew tensor to self.step
-                self.data[prl.REW][self.step][i].copy_(self.demos_in_progress["env{}".format(
-                    i + 1)]["Demo"][prl.REW][demo_step])
+                self.sample_demo(env_id=i)
+
+        if len(all_envs) > 0:
+
+            # Cast obs, rhs, dones into the right format
+            all_obs = torch.cat(all_obs)
+            all_done = torch.cat(all_done)
+            for k in all_rhs:
+                all_rhs[k] = torch.cat(all_rhs[k])
+
+            # Run forward pass
+            _, _, rhs2, algo_data = self.algo.acting_step(all_obs, all_rhs, all_done)
+
+            for num, i in enumerate(all_envs):
+
+                demo_step = self.demos_in_progress["env{}".format(i + 1)]["Step"]
+
+                # Insert demo act and rew tensors to self.step
+                for tensor in (prl.ACT, prl.REW):
+                    self.data[tensor][self.step][i].copy_(
+                        torch.FloatTensor(self.demos_in_progress["env{}".format(i + 1)]["Demo"][tensor][demo_step]))
 
                 # Insert demo logprob to self.step. Demo action prob is 1.0, so logprob is 0.0
                 self.data[prl.LOGP][self.step][i].copy_(torch.zeros(1))
 
-                # Insert demo values predicted by the forward pass
-                self.data[prl.VAL][self.step][i].copy_(algo_data[prl.VAL].squeeze())
+                # Insert other tensors predicted in the forward pass
+                for tensor in (prl.IREW, prl.VAL, prl.IVAL):
+                    if tensor in algo_data.keys():
+                        self.data[tensor][self.step][i].copy_(algo_data[tensor][num])
 
                 # Update demo_in_progress variables
                 self.demos_in_progress["env{}".format(i + 1)]["Step"] += 1
-                self.demos_in_progress["env{}".format(i + 1)][prl.RHS] = rhs2
+
+                self.demos_in_progress["env{}".format(i + 1)][prl.RHS] = {
+                    k: rhs2[k][num].reshape(1, -1) for k in rhs2.keys()}
 
                 self.demos_in_progress["env{}".format(i + 1)]["MaxValue"] = max(
-                    [algo_data[prl.VAL].item(), self.demos_in_progress["env{}".format(i + 1)]["MaxValue"]])
+                    [algo_data[prl.VAL][num].item(), self.demos_in_progress["env{}".format(i + 1)]["MaxValue"]])
+
+                self.inserted_samples += 1
 
                 # Handle end of demos
                 if demo_step == self.demos_in_progress["env{}".format(i + 1)]["DemoLength"] - 1:
@@ -332,20 +404,15 @@ class PPODBuffer(B):
                     self.data[prl.DONE][self.step + 1][i].copy_(torch.zeros(1))
 
                     # Insert demo obs2 tensor to self.step + 1
-                    obs2 = torch.roll(obs, -self.num_channels_obs, dims=1).squeeze(0)
-                    obs2[-self.num_channels_obs:].copy_(
-                        self.demos_in_progress["env{}".format(i + 1)]["Demo"][prl.OBS][demo_step + 1].to(self.device))
+                    obs2 = torch.roll(all_obs[num:num+1], -self.num_channels_obs, dims=1).squeeze(0)
+                    obs2[-self.num_channels_obs:].copy_(torch.FloatTensor(
+                        self.demos_in_progress["env{}".format(i + 1)]["Demo"][prl.OBS][demo_step + 1]))
                     self.data[prl.OBS][self.step + 1][i].copy_(obs2)
 
                     # Insert demo rhs2 tensor to self.step + 1
                     if self.recurrent_actor:
                         for k in self.data[prl.RHS]:
-                            self.data[prl.RHS][k][self.step + 1][i].copy_(rhs2[k].squeeze())
-
-            # Otherwise check if end of episode reached and randomly start new demo
-            elif sample[prl.DONE2][i] == 1.0:
-
-                self.sample_demo(env_id=i)
+                            self.data[prl.RHS][k][self.step + 1][i].copy_(rhs2[k][num].squeeze())
 
         self.step = (self.step + 1) % self.max_size
         self.size = min(self.size + 1, self.max_size)
@@ -358,11 +425,11 @@ class PPODBuffer(B):
             # Copy transition
             for tensor in self.demos_data_fields:
                 if tensor in (prl.OBS):
-                    self.potential_demos["env{}".format(i + 1)][tensor].append(
-                        copy.deepcopy(sample[tensor][i, -self.num_channels_obs:]).cpu().numpy())
+                    self.potential_demos["env{}".format(i + 1)][tensor].append(copy.deepcopy(
+                        sample[tensor][i, -self.num_channels_obs:]).cpu().numpy().astype(self.demo_dtypes[tensor]))
                 else:
                     self.potential_demos["env{}".format(i + 1)][tensor].append(
-                        copy.deepcopy(sample[tensor][i]).cpu().numpy())
+                        copy.deepcopy(sample[tensor][i]).cpu().numpy().astype(self.demo_dtypes[tensor]))
 
             # Track highest value prediction
             self.potential_demos_val[i] = max([self.potential_demos_val["env{}".format(
@@ -374,17 +441,29 @@ class PPODBuffer(B):
                 # Get candidate demos
                 potential_demo = {}
                 for tensor in self.demos_data_fields:
-                    potential_demo[tensor] = torch.Tensor(np.stack(
-                        self.potential_demos["env{}".format(i + 1)][tensor]))
+                    potential_demo[tensor] = np.stack(self.potential_demos["env{}".format(i + 1)][tensor])
+                    if tensor == prl.REW:
+                        nonzero = np.flatnonzero(potential_demo[tensor] > 0.0)
+                        last_reward = len(potential_demo[tensor]) if len(nonzero) == 0 else np.max(nonzero)
+
+                print(last_reward)
+
+                # Cut off data after last reward
+                for tensor in self.demos_data_fields:
+                    potential_demo[tensor] = potential_demo[tensor][0:last_reward + 1]
 
                 # Compute accumulated reward
-                episode_reward = potential_demo[prl.REW].sum().item()
+                episode_reward = potential_demo[prl.REW].sum()
                 potential_demo["ID"] = str(uuid.uuid4())
                 potential_demo["TotalReward"] = episode_reward
                 potential_demo["DemoLength"] = potential_demo[prl.ACT].shape[0]
 
                 # Consider candidate demos for demos reward
                 if episode_reward >= self.reward_threshold:
+
+                    # If better than any other existing demo, empty buffer
+                    if episode_reward > self.max_demo_reward:
+                        self.reward_demos = []
 
                     # Add demos to reward buffer
                     self.reward_demos.append(potential_demo)
@@ -450,25 +529,15 @@ class PPODBuffer(B):
                     raise ValueError(
                         "Env and demo with different frame skip!")
 
-                # Add action
-                demo_act = demo[prl.ACT]
-                self.demo_act_dtype = demo_act.dtype
-                new_demo[prl.ACT] = torch.FloatTensor(demo_act)
-
-                # Add obs
-                demo_obs = demo[prl.OBS]
-                self.demo_obs_dtype = demo_obs.dtype
-                new_demo[prl.OBS] = torch.FloatTensor(demo_obs)
-
-                # Add rew
-                demo_rew = demo[prl.REW]
-                self.demo_rew_dtype = demo_rew.dtype
-                new_demo[prl.REW] = torch.FloatTensor(demo_rew)
+                # Add action, obs, rew
+                new_demo[prl.ACT] = demo[prl.ACT]
+                new_demo[prl.OBS] = demo[prl.OBS]
+                new_demo[prl.REW] = demo[prl.REW]
 
                 new_demo.update({
                     "ID": str(uuid.uuid4()),
                     "DemoLength": demo[prl.ACT].shape[0],
-                    "TotalReward": demo_rew.sum().item()})
+                    "TotalReward": new_demo[prl.REW].sum()})
                 self.reward_demos.append(new_demo)
                 num_loaded_human_demos += 1
 
@@ -483,29 +552,19 @@ class PPODBuffer(B):
                 demo = np.load(demo_file)
                 new_demo = {k: {} for k in self.demos_data_fields}
 
-                if demo["FrameSkip"] != self.frame_skip:
+                if int(demo["FrameSkip"]) != self.frame_skip:
                     raise ValueError(
                         "Env and demo with different frame skip!")
 
-                # Add action
-                demo_act = demo[prl.ACT]
-                self.demo_act_dtype = demo_act.dtype
-                new_demo[prl.ACT] = torch.FloatTensor(demo_act)
-
-                # Add obs
-                demo_obs = demo[prl.OBS]
-                self.demo_obs_dtype = demo_obs.dtype
-                new_demo[prl.OBS] = torch.FloatTensor(demo_obs)
-
-                # Add rew
-                demo_rew = demo[prl.REW]
-                self.demo_rew_dtype = demo_rew.dtype
-                new_demo[prl.REW] = torch.FloatTensor(demo_rew)
+                # Add action, obs, rew
+                new_demo[prl.ACT] = demo[prl.ACT]
+                new_demo[prl.OBS] = demo[prl.OBS]
+                new_demo[prl.REW] = demo[prl.REW]
 
                 new_demo.update({
                     "ID": str(uuid.uuid4()),
                     "DemoLength": demo[prl.ACT].shape[0],
-                    "TotalReward": demo_rew.sum().item()})
+                    "TotalReward": new_demo[prl.REW].sum()})
                 self.reward_demos.append(new_demo)
                 num_loaded_reward_demos += 1
 
@@ -566,8 +625,8 @@ class PPODBuffer(B):
             # Set next buffer obs to be the starting demo obs
             for k in range(self.frame_stack):
                 self.data[prl.OBS][self.step + 1][env_id][
-                k * self.num_channels_obs:(k + 1) * self.num_channels_obs].copy_(
-                    self.demos_in_progress["env{}".format(env_id + 1)]["Demo"][prl.OBS][0].to(self.device))
+                k * self.num_channels_obs:(k + 1) * self.num_channels_obs].copy_(torch.FloatTensor(
+                    self.demos_in_progress["env{}".format(env_id + 1)]["Demo"][prl.OBS][0]))
 
         else:
             # Reset `i-th` environment as set next buffer obs to be the starting episode obs
@@ -599,15 +658,16 @@ class PPODBuffer(B):
         total_demos = len(self.reward_demos) + len(self.value_demos)
         if total_demos > self.max_demos:
             for _ in range(min(total_demos - self.max_demos, len(self.value_demos))):
+
                 # Pop value demos with lowest MaxValue
                 del self.value_demos[np.array([p["MaxValue"] for p in self.value_demos]).argmin()]
 
         # If after popping all value demos, still over max_demos, pop reward demos
         if len(self.reward_demos) > self.max_demos:
             for _ in range(len(self.reward_demos) - self.max_demos):
-                # Eject agent demo with lowest reward
-                rewards = np.array([p[prl.REW].sum() for p in self.reward_demos[self.num_loaded_human_demos:]])
-                del self.reward_demos[np.argmin(rewards) + self.num_loaded_human_demos]
+                # Option 2: Eject longer agent demo
+                lengths = np.array([p[prl.OBS].shape[0] for p in self.reward_demos[self.num_loaded_human_demos:]])
+                del self.reward_demos[np.argmax(lengths) + self.num_loaded_human_demos]
 
     def save_demos(self):
         """
@@ -632,7 +692,7 @@ class PPODBuffer(B):
                 demo_pos += self.num_loaded_human_demos
                 np.savez(
                     os.path.join(self.target_agent_demos_dir, filename),
-                    Observation=np.array(self.reward_demos[demo_pos][prl.OBS]).astype(self.demo_obs_dtype),
-                    Reward=np.array(self.reward_demos[demo_pos][prl.REW]).astype(self.demo_rew_dtype),
-                    Action=np.array(self.reward_demos[demo_pos][prl.ACT]).astype(self.demo_act_dtype),
+                    Observation=np.array(self.reward_demos[demo_pos][prl.OBS]).astype(self.demo_dtypes[prl.OBS]),
+                    Reward=np.array(self.reward_demos[demo_pos][prl.REW]).astype(self.demo_dtypes[prl.REW]),
+                    Action=np.array(self.reward_demos[demo_pos][prl.ACT]).astype(self.demo_dtypes[prl.ACT]),
                     FrameSkip=self.frame_skip)
